pyproject.toml
[project]
name = "simple-plain-coder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.9.1",
    "platformdirs>=4.2.0",
    "prompt-toolkit>=3.0.51",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "qwen-agent>=0.0.25",
    "tomli>=2.2.1",
    "tomli-w>=1.0.0",
]
